refactor(peagen): move schemas module
